Enhancement was made to the content description sub function thanks to sonntux submission on git hub.
There were an issue where a double space could occure, when the following
conditions were meet. When at the end of a line with a special HTML char
and following bracketed content, the replacement algorithm produced a new
space char. In combination with the usual following space a new double empty
space have been created. A correction double empty space was already implemented
but was applied before the occurence of this newly created double empty spaces

The programm flow was adjusted to take care of the issue.